IOS: Move USB scanning thread to new class USBScanner
