+++ Source/Core/Core/IOS/USB/Host.cpp
@@ -3,30 +3,17 @@
 
 #include "Core/IOS/USB/Host.h"
 
-#include <algorithm>
+#include <functional>
 #include <memory>
-#include <mutex>
-#include <set>
+#include <optional>
 #include <string>
-#include <utility>
 
-#ifdef __LIBUSB__
-#include <libusb.h>
-#endif
-
-#include "Common/Assert.h"
 #include "Common/ChunkFile.h"
 #include "Common/CommonTypes.h"
 #include "Common/Logging/Log.h"
-#include "Common/Thread.h"
-#include "Core/Config/MainSettings.h"
 #include "Core/Core.h"
 #include "Core/IOS/USB/Common.h"
-#include "Core/IOS/USB/Emulated/Infinity.h"
-#include "Core/IOS/USB/Emulated/Skylanders/Skylander.h"
-#include "Core/IOS/USB/LibusbDevice.h"
-#include "Core/NetPlayProto.h"
-#include "Core/System.h"
+#include "Core/IOS/USB/USBScanner.h"
 
 namespace IOS::HLE
 {
@@ -35,16 +22,19 @@ USBHost::USBHost(EmulationKernel& ios, const std::string& device_name)
 {
 }
 
-USBHost::~USBHost() = default;
+USBHost::~USBHost()
+{
+  m_usb_scanner.Stop();
+}
 
 std::optional<IPCReply> USBHost::Open(const OpenRequest& request)
 {
   if (!m_has_initialised)
   {
-    GetScanThread().Start();
+    m_usb_scanner.Start();
     // Force a device scan to complete, because some games (including Your Shape) only care
     // about the initial device list (in the first GETDEVICECHANGE reply).
-    GetScanThread().WaitForFirstScan();
+    m_usb_scanner.WaitForFirstScan();
     m_has_initialised = true;
   }
   return IPCReply(IPC_SUCCESS);
@@ -53,9 +43,9 @@ std::optional<IPCReply> USBHost::Open(const OpenRequest& request)
 void USBHost::UpdateWantDeterminism(const bool new_want_determinism)
 {
   if (new_want_determinism)
-    GetScanThread().Stop();
+    m_usb_scanner.Stop();
   else if (IsOpened())
-    GetScanThread().Start();
+    m_usb_scanner.Start();
 }
 
 void USBHost::DoState(PointerWrap& p)
@@ -65,27 +55,13 @@ void USBHost::DoState(PointerWrap& p)
   {
     // After a state has loaded, there may be insertion hooks for devices that were
     // already plugged in, and which need to be triggered.
-    UpdateDevices(true);
+    m_usb_scanner.UpdateDevices(true);
   }
-}
-
-bool USBHost::AddDevice(std::unique_ptr<USB::Device> device)
-{
-  std::lock_guard lk(m_devices_mutex);
-  if (m_devices.contains(device->GetId()))
-    return false;
-
-  m_devices[device->GetId()] = std::move(device);
-  return true;
 }
 
 std::shared_ptr<USB::Device> USBHost::GetDeviceById(const u64 device_id) const
 {
-  std::lock_guard lk(m_devices_mutex);
-  const auto it = m_devices.find(device_id);
-  if (it == m_devices.end())
-    return nullptr;
-  return it->second;
+  return m_usb_scanner.GetDeviceById(device_id);
 }
 
 void USBHost::OnDeviceChange(ChangeEvent event, std::shared_ptr<USB::Device> changed_device)
@@ -104,69 +80,7 @@ bool USBHost::ShouldAddDevice(const USB::Device& device) const
 void USBHost::Update()
 {
   if (Core::WantsDeterminism())
-    UpdateDevices();
-}
-
-// This is called from the scan thread. Returns false if we failed to update the device list.
-bool USBHost::UpdateDevices(const bool always_add_hooks)
-{
-  DeviceChangeHooks hooks;
-  std::set<u64> plugged_devices;
-  // If we failed to get a new, up-to-date list of devices, we cannot detect device removals.
-  if (!AddNewDevices(plugged_devices, hooks, always_add_hooks))
-    return false;
-  DetectRemovedDevices(plugged_devices, hooks);
-  DispatchHooks(hooks);
-  return true;
-}
-
-bool USBHost::AddNewDevices(std::set<u64>& new_devices, DeviceChangeHooks& hooks,
-                            const bool always_add_hooks)
-{
-  AddEmulatedDevices(new_devices, hooks, always_add_hooks);
-#ifdef __LIBUSB__
-  if (!Core::WantsDeterminism())
-  {
-    auto whitelist = Config::GetUSBDeviceWhitelist();
-    if (whitelist.empty())
-      return true;
-
-    if (m_context.IsValid())
-    {
-      const int ret = m_context.GetDeviceList([&](libusb_device* device) {
-        libusb_device_descriptor descriptor;
-        libusb_get_device_descriptor(device, &descriptor);
-        if (!whitelist.contains({descriptor.idVendor, descriptor.idProduct}))
-          return true;
-
-        auto usb_device =
-            std::make_unique<USB::LibusbDevice>(GetEmulationKernel(), device, descriptor);
-        CheckAndAddDevice(std::move(usb_device), new_devices, hooks, always_add_hooks);
-        return true;
-      });
-      if (ret != LIBUSB_SUCCESS)
-        WARN_LOG_FMT(IOS_USB, "GetDeviceList failed: {}", LibusbUtils::ErrorWrap(ret));
-    }
-  }
-#endif
-  return true;
-}
-
-void USBHost::DetectRemovedDevices(const std::set<u64>& plugged_devices, DeviceChangeHooks& hooks)
-{
-  std::lock_guard lk(m_devices_mutex);
-  for (auto it = m_devices.begin(); it != m_devices.end();)
-  {
-    if (!plugged_devices.contains(it->second->GetId()))
-    {
-      hooks.emplace(it->second, ChangeEvent::Removed);
-      it = m_devices.erase(it);
-    }
-    else
-    {
-      ++it;
-    }
-  }
+    m_usb_scanner.UpdateDevices();
 }
 
 void USBHost::DispatchHooks(const DeviceChangeHooks& hooks)
@@ -180,80 +94,6 @@ void USBHost::DispatchHooks(const DeviceChangeHooks& hooks)
   }
   if (!hooks.empty())
     OnDeviceChangeEnd();
-}
-
-void USBHost::AddEmulatedDevices(std::set<u64>& new_devices, DeviceChangeHooks& hooks,
-                                 bool always_add_hooks)
-{
-  if (Config::Get(Config::MAIN_EMULATE_SKYLANDER_PORTAL) && !NetPlay::IsNetPlayRunning())
-  {
-    auto skylanderportal = std::make_unique<USB::SkylanderUSB>(GetEmulationKernel());
-    CheckAndAddDevice(std::move(skylanderportal), new_devices, hooks, always_add_hooks);
-  }
-  if (Config::Get(Config::MAIN_EMULATE_INFINITY_BASE) && !NetPlay::IsNetPlayRunning())
-  {
-    auto infinity_base = std::make_unique<USB::InfinityUSB>(GetEmulationKernel());
-    CheckAndAddDevice(std::move(infinity_base), new_devices, hooks, always_add_hooks);
-  }
-}
-
-void USBHost::CheckAndAddDevice(std::unique_ptr<USB::Device> device, std::set<u64>& new_devices,
-                                DeviceChangeHooks& hooks, bool always_add_hooks)
-{
-  if (ShouldAddDevice(*device))
-  {
-    const u64 deviceid = device->GetId();
-    new_devices.insert(deviceid);
-    if (AddDevice(std::move(device)) || always_add_hooks)
-    {
-      hooks.emplace(GetDeviceById(deviceid), ChangeEvent::Inserted);
-    }
-  }
-}
-
-USBHost::ScanThread::~ScanThread()
-{
-  Stop();
-}
-
-void USBHost::ScanThread::WaitForFirstScan()
-{
-  if (m_thread_running.IsSet())
-  {
-    m_first_scan_complete_event.Wait();
-  }
-}
-
-void USBHost::ScanThread::Start()
-{
-  if (Core::WantsDeterminism())
-  {
-    m_host->UpdateDevices();
-    return;
-  }
-  if (m_thread_running.TestAndSet())
-  {
-    m_thread = std::thread([this] {
-      Common::SetCurrentThreadName("USB Scan Thread");
-      while (m_thread_running.IsSet())
-      {
-        if (m_host->UpdateDevices())
-          m_first_scan_complete_event.Set();
-        Common::SleepCurrentThread(50);
-      }
-    });
-  }
-}
-
-void USBHost::ScanThread::Stop()
-{
-  if (m_thread_running.TestAndClear())
-    m_thread.join();
-
-  // Clear all devices and dispatch removal hooks.
-  DeviceChangeHooks hooks;
-  m_host->DetectRemovedDevices(std::set<u64>(), hooks);
-  m_host->DispatchHooks(hooks);
 }
 
 std::optional<IPCReply> USBHost::HandleTransfer(std::shared_ptr<USB::Device> device, u32 request,
@@ -270,4 +110,5 @@ std::optional<IPCReply> USBHost::HandleTransfer(std::shared_ptr<USB::Device> dev
                 device->GetVid(), device->GetPid(), request, device->GetErrorName(ret));
   return IPCReply(ret <= 0 ? ret : IPC_EINVAL);
 }
+
 }  // namespace IOS::HLE
