+++ Source/Core/Core/IOS/USB/USB_HID/HIDv5.cpp
@@ -18,10 +18,7 @@ namespace IOS::HLE
 {
 constexpr u32 USBV5_VERSION = 0x50001;
 
-USB_HIDv5::~USB_HIDv5()
-{
-  m_scan_thread.Stop();
-}
+USB_HIDv5::~USB_HIDv5() = default;
 
 std::optional<IPCReply> USB_HIDv5::IOCtl(const IOCtlRequest& request)
 {
