+++ Source/Core/Core/IOS/USB/USB_VEN/VEN.cpp
@@ -18,10 +18,7 @@ namespace IOS::HLE
 {
 constexpr u32 USBV5_VERSION = 0x50001;
 
-USB_VEN::~USB_VEN()
-{
-  m_scan_thread.Stop();
-}
+USB_VEN::~USB_VEN() = default;
 
 std::optional<IPCReply> USB_VEN::IOCtl(const IOCtlRequest& request)
 {
